+++ .pre-commit-config.yaml
@@ -9,6 +9,7 @@ repos:
     hooks:
       - id: black
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v1.4.0
+    rev: v1.4.0-1
     hooks:
       - id: flake8
+        additional_dependencies: [flake8-bugbear]
