+++ setup.cfg
@@ -32,3 +32,5 @@ select = B, E, F, W, B9
 ignore = E203, E501, W503
 # up to 88 allowed by bugbear B950
 max-line-length = 80
+# init is used to export public API, ignore import warnings
+exclude = src/itsdangerous/__init__.py
